fix: add feedback to purchases

+++ frapp/src/modules/purchases/components/ItemsProducts.tsx
@@ -1,29 +1,39 @@
 import {
   Box,
   Button,
+  Divider,
   FormControl,
-  FormLabel,
   HStack,
+  Stat,
+  StatLabel,
+  StatNumber,
   VStack,
 } from "@chakra-ui/react";
 import { atom, useSetAtom } from "jotai";
 import { FC } from "react";
-import { DeleteIcon, EditIcon, List, ListItem, item } from "~/ui";
+import {
+  DeleteIcon,
+  EditIcon,
+  List,
+  ListItem,
+  item,
+  moneyStrategyFormat,
+} from "~/ui";
 import { makeDisclosure } from "~/utils";
 import { LineSale, saleItemAtom } from "../data";
-import { useHandleLineSale } from "../helpers/useHandleLineSale";
-import { MyDrawer } from "./MyDrawerItem";
 import {
   ItemsProductContextProps,
   ItemsProductProvider,
+  useItemsProductContext,
 } from "../helpers/ItemsProductContext";
-import { useItemsProductContext } from "../helpers/ItemsProductContext";
-
+import { useHandleLineSale } from "../helpers/useHandleLineSale";
+import { MyDrawer } from "./MyDrawerItem";
+import { useConfirmDialog } from "~/ui";
 const lineSaleAtomDrawer = atom(false);
 
 export const useSalelineDisclosure = makeDisclosure(lineSaleAtomDrawer);
 
-const LineSaleItem: FC<{
+const LinePurchaseItem: FC<{
   line: LineSale;
 }> = ({ line }) => {
   const drawerState = useSalelineDisclosure();
@@ -33,22 +43,30 @@ const LineSaleItem: FC<{
 
   const itemsProducts = useItemsProductContext();
 
+  const confirmDialog = useConfirmDialog();
+
   const actions = itemsProducts.isEdit ? (
     <>
       <Button
         onClick={() => {
           drawerState.onOpen();
+
           setSaleItem(line);
         }}
         colorScheme="blue"
       >
         <EditIcon />
       </Button>
-
       <Button
         colorScheme="red"
         onClick={() => {
-          deleteLine(line.productId);
+          confirmDialog.open({
+            title: "Eliminar",
+            description: "¿Está seguro de eliminar este item?",
+            onConfirm: () => {
+              deleteLine(line.productId);
+            },
+          });
         }}
       >
         <DeleteIcon />
@@ -61,9 +79,9 @@ const LineSaleItem: FC<{
       label={
         <VStack alignItems={"flex-start"}>
           {item("Producto", line.productName)}
-          {item("Total", line.total)}
-          {item("Cantidad", line.amount)}
-          {item("Precio", line.price)}
+          {item("Total", moneyStrategyFormat.format(line.total))}
+          {item("Cantidad", moneyStrategyFormat.format(line.amount))}
+          {item("Precio", moneyStrategyFormat.format(line.price))}
         </VStack>
       }
       actions={actions}
@@ -72,11 +90,18 @@ const LineSaleItem: FC<{
 };
 
 export const ItemsProducts: FC<Partial<ItemsProductContextProps>> = ({
-  isEdit = false,
+  isEdit = true,
 }) => {
   const drawerState = useSalelineDisclosure();
-  const { lines } = useHandleLineSale();
+  const { lines, getTotal } = useHandleLineSale();
   const setSaleItem = useSetAtom(saleItemAtom);
+
+  const moneyElement = (
+    <Stat>
+      <StatLabel>Total:</StatLabel>
+      <StatNumber>{moneyStrategyFormat.format(getTotal())}</StatNumber>
+    </Stat>
+  );
 
   return (
     <ItemsProductProvider
@@ -84,12 +109,17 @@ export const ItemsProducts: FC<Partial<ItemsProductContextProps>> = ({
         isEdit,
       }}
     >
-      <Box mt={4}>
+      <Box mt={4} borderWidth={"1px"} borderColor={"gray.200"} p="2">
         <FormControl>
-          <FormLabel>Items</FormLabel>
           <Box>
             {!isEdit && (
-              <HStack justifyContent={"flex-end"}>
+              <HStack justifyContent={"space-between"} px="2">
+                {moneyElement}
+              </HStack>
+            )}
+            {isEdit && (
+              <HStack justifyContent={"space-between"}>
+                {moneyElement}
                 <Button
                   onClick={() => {
                     drawerState.onOpen();
@@ -100,10 +130,11 @@ export const ItemsProducts: FC<Partial<ItemsProductContextProps>> = ({
                 </Button>
               </HStack>
             )}
+            <Divider borderWidth={"1px"} mt="2" borderColor={"gray.300"} />
             <List
               data={lines}
               renderItem={(line) => {
-                return <LineSaleItem line={line} />;
+                return <LinePurchaseItem line={line} />;
               }}
             />
           </Box>
